Rename read module to bufread
